services/api/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ── Database (async) ──────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0

# ── Migrations (also run from CI — not the API process itself) ────────────────
alembic>=1.13.0
psycopg2-binary>=2.9.9   # alembic env.py uses sync driver for DDL

# ── Data validation ───────────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ── Task queue (client only — no worker process in this service) ──────────────
celery[redis]>=5.4.0
redis>=5.0.0

# ── HTTP client (Ollama LLM + embedding calls) ────────────────────────────────
httpx>=0.27.0

# ── Vector store ──────────────────────────────────────────────────────────────
weaviate-client>=4.6.0

# ── PDF text extraction (for /v1/documents/upload-pdf endpoint) ───────────────
# PyMuPDF ≥ 1.23 ships pre-compiled wheels — no system MuPDF required.
pymupdf>=1.24.0

# ── Multipart form uploads (PDF file upload support) ─────────────────────────
python-multipart>=0.0.9
